src/requirements.txt
# watchlist
publicsuffix

# database
elasticsearch
requests

# whois enrichment
pythonwhois

# phishing score
entropy==0.10
tqdm==4.19.4
tld==0.7.9
python_Levenshtein==0.12.0
PyYAML==3.13

# certstream
certstream==1.10
websocket-client==0.48.0